Improve HTTPS and HTTP Server
**NEWHTTPSServer.java, NEWHTTPServer.java, NEWHTTPServerCommon.java**
- Moved the core functions from NEWHTTPServer and NEWHTTPSServer to NEWHTTPServerCommon

**YTWebSocketSecureServer.java**
- Added SSL support for YTP websocket server

**PhantomBot.java**
- Error when usehttps is enabled and no keystore is provided
- Load the YTWebSocketSecureServer when required

**ytPlayer.js**
- Support YTWebSocketSecureServer
